requirements.txt
flask
qrcode
pyshorteners
requests
pytesseract
Pillow
cs50
